Module-4_VoiceRecognition_ConversationalEngine/voice_engine/requirements.txt
requests>=2.31.0
sounddevice>=0.4.6
numpy>=1.24.0
pyttsx3>=2.90
pygame>=2.5.0
RPi.GPIO>=0.7.1
google-generativeai>=0.3.0